CE: Make HCE able to handle long APDU (>250 bytes)
BZ: 163352

Due to a bug in the type error  in the prototype of
RoutingManager::handle() function Android cannot handles
APDUs which have size above 255 bytes.
In order to be able to handle longer APDUs, the size has  to be
stored within a UINT16.

Category: device enablement
Domain: CWS.NFC-CE
Origin: internal
Upstream-Candidate: yes

+++ nci/jni/RoutingManager.cpp
@@ -432,7 +432,7 @@ void RoutingManager::notifyDeactivated ()
     }
 }
 
-void RoutingManager::handleData (const UINT8* data, UINT8 dataLen)
+void RoutingManager::handleData (const UINT8* data, UINT16 dataLen)
 {
     if (dataLen <= 0)
     {
